+++ magpie/api/management/user/user_utils.py
@@ -34,7 +34,7 @@ from magpie.api.webhooks import (
     process_webhook_requests
 )
 from magpie.constants import get_constant
-from magpie.models import TemporaryToken, TokenOperation
+from magpie.models import TemporaryToken, TokenOperation, UserGroupStatus
 from magpie.permissions import PermissionSet, PermissionType, format_permissions
 from magpie.services import SERVICE_TYPE_DICT, service_factory
 from magpie.ui.utils import BaseViews
@@ -871,6 +871,27 @@ def check_user_info(user_name=None, email=None, password=None, group_name=None, 
         ax.verify_param(group_name, matches=True, param_name="group_name", param_compare=ax.PARAM_REGEX,
                         http_error=HTTPBadRequest,
                         msg_on_fail=s.Users_CheckInfo_GroupName_BadRequestResponseSchema.description)
+
+
+def get_user_groups(user, status, db_session):
+    # type: (models.User, models.UserGroupStatus, Session) -> List[Str]
+    """
+    List all groups a user belongs to, filtered by UserGroup status type.
+    """
+    group_names = set()
+    member_group_names = set(get_user_groups_checked(user, db_session))
+    if status in [UserGroupStatus.ACTIVE, UserGroupStatus.ALL]:
+        group_names = group_names.union(member_group_names)
+    if status in [UserGroupStatus.PENDING, UserGroupStatus.ALL]:
+        tmp_tokens = TemporaryToken.by_user(user).filter(
+            TemporaryToken.operation == TokenOperation.GROUP_ACCEPT_TERMS)
+        pending_group_names = set(tmp_token.group.group_name for tmp_token in tmp_tokens)
+
+        # Remove any group a user already belongs to, in case any tokens are irrelevant.
+        # Should not happen since related tokens are deleted upon T&C acceptation.
+        pending_group_names = pending_group_names - member_group_names
+        group_names = group_names.union(pending_group_names)
+    return group_names
 
 
 def get_user_groups_checked(user, db_session):